refactor(chat): extract ChatInputField into dedicated file and fix send-trim inconsistency
- Move ChatInputFieldProps, ChatInputField, and SendButton from ChatComponents.kt
  into new ChatInputField.kt to resolve Detekt TooManyFunctions violation
- Fix send path to dispatch props.value unchanged; trim is now only used for
  the isEnabled guard, keeping intentional whitespace in the payload
- Guard sendMessage() with prompt.trim().isBlank() to stay consistent with UI
- Add missing blank line after schedule heading in automation.md (MD022)
